+++ docs/changes.rst
@@ -4,7 +4,7 @@ Wand Changelog
 Version 0.1.5
 -------------
 
-To be released. Slightly mature alpha version.
+Released on October 28, 2011. Slightly mature alpha version.
 
  - Now :class:`~wand.image.Image` can read Python file objects by ``file``
    keyword argument.